feat: refactor password change and account deletion logic in profile
